Create Challenge_3.Rmd
The main problem here is I could not get the urban area data. It was actually working at one point which is why I have functions below it but I changed something while organizing and it no longer works.
It will work for a single continent but when I try to pass the whole column into the function it breaks.
Any advice would be great because I can see how GET is powerful and I would like to know how to do this even though the assignment has passed.